rating: widen rated_usage.event_count INTEGER → BIGINT
The token-sum and count casts in the rater were already ::bigint, but the
rated_usage.event_count COLUMN was still INTEGER and the per-rollup
COUNT(*)::int cast still narrowed — an asymmetry that would fail the INSERT
on a rollup counting more than 2^31 events. Widen the column (both the raw
SQL migration and the Atlas mirror) and the COUNT(*) cast to bigint so the
whole path is int64-clean (EventsRated was already int64).

No data migration: rated_usage is empty pre-prod, so this is a plain
column-type change with no backfill.

+++ internal/rating/store.go
@@ -325,7 +325,7 @@ priced AS (
           + cached_tokens     * cached_price
           + completion_tokens * completion_price
         )                                                AS cost,
-        COUNT(*)::int                                    AS event_count
+        COUNT(*)::bigint                                 AS event_count
     FROM resolved
     WHERE prompt_price IS NOT NULL          -- priced only
       AND auth_id  IS NOT NULL              -- attributable only
